1-created a class for Conjured Items, 2- moved classes to protocols folder
